move evaluatePathExpressions and evaluateScalarTokens into graphite.render.evaluator
